refactor: moving data types to types.rs
